pyproject.toml
[project]
name = "read-pdf-agent"
version = "1.0.0"
description = "A2A Document Extraction Agent for PDFs"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "pdfplumber>=0.11.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "a2a-sdk>=0.2.11",
    "uvicorn>=0.34.3",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"